+++ src/main/java/com/assignment/rewardmanagement/dto/response/CustomerResponse.java
@@ -2,24 +2,24 @@ package com.assignment.rewardmanagement.dto.response;
 
 public class CustomerResponse {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
 
     public CustomerResponse() {
     }
 
-    public CustomerResponse(Integer id, String name, String email) {
+    public CustomerResponse(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
