+++ src/main/java/com/assignment/rewardmanagement/service/RewardService.java
@@ -7,9 +7,9 @@ import com.assignment.rewardmanagement.dto.response.RewardResponse;
 
 public interface RewardService {
 
-    RewardResponse getRewardPointsForCustomer(Integer customerId);
+    RewardResponse getRewardPointsForCustomer(Long customerId);
 
-    RewardResponse getRewardPointsForCustomerInRange(Integer customerId, LocalDateTime start, LocalDateTime end);
+    RewardResponse getRewardPointsForCustomerInRange(Long customerId, LocalDateTime start, LocalDateTime end);
 
     List<RewardResponse> getAllCustomerRewards();
 }
